+++ server/src/main/java/org/allaymc/server/block/component/ice/BlockIceBaseComponentImpl.java
@@ -9,7 +9,9 @@ import org.allaymc.api.entity.Entity;
 import org.allaymc.api.eventbus.event.block.BlockFadeEvent;
 import org.allaymc.api.item.ItemStack;
 import org.allaymc.api.utils.Utils;
+import org.allaymc.api.world.Dimension;
 import org.allaymc.server.block.component.BlockBaseComponentImpl;
+import org.joml.Vector3ic;
 
 import java.util.Set;
 
@@ -25,13 +27,24 @@ public class BlockIceBaseComponentImpl extends BlockBaseComponentImpl {
     public void onRandomUpdate(BlockStateWithPos current) {
         super.onRandomUpdate(current);
 
-        if (current.getDimension().getLightService().getInternalLight(current.getPos()) > 11) {
+        if (isNeighborBlockLightBiggerThan(current.dimension(), current.pos(), 11)) {
             var event = new BlockFadeEvent(current, BlockTypes.WATER.getDefaultState());
             if (event.call()) {
                 current.getDimension().setBlockState(current.getPos(), event.getNewBlockState());
             }
         }
     }
+
+    protected boolean isNeighborBlockLightBiggerThan(Dimension dimension, Vector3ic pos, int value) {
+        for (var face : BlockFace.values()) {
+            if (dimension.getLightService().getBlockLight(face.offsetPos(pos)) > value) {
+                return true;
+            }
+        }
+
+        return false;
+    }
+
 
     @Override
     public Set<ItemStack> getDrops(BlockStateWithPos current, ItemStack usedItem, Entity entity) {