+++ server/src/main/java/org/allaymc/server/world/generator/AllayWorldGenerator.java
@@ -92,7 +92,7 @@ public class AllayWorldGenerator implements WorldGenerator {
     }
 
     public void startTick() {
-        Thread.ofVirtual().name("Population Queue Processing Thread - " + dimension.getWorld().getWorldData().getDisplayName()).start(() -> {
+        Thread.ofVirtual().name("Population Queue Processing Thread - " + dimension.getWorld().getName()).start(() -> {
             while (dimension.getWorld().isRunning()) {
                 processPopulationQueue();
             }